rename and resolve references
